siam_1912022_2

+++ Core/Entities/LastFiveInvestmentInfo.cs
@@ -4,7 +4,7 @@ namespace Core.Entities
     public class LastFiveInvestmentInfo : BaseEntity
     {
         public string? DonationShortName { get; set; }
-        public long? InvestmentAmount { get; set; }
+        public double? InvestmentAmount { get; set; }
         public string? ComtSharePrcntAll { get; set; }
         public string? ComtSharePrcnt { get; set; }
         public double? PrescribedSharePrcnt { get; set; }
